% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{out_list_demo}
\alias{out_list_demo}
\title{Out demo}
\format{
A data.table data frame with X rows and Y variables:
\describe{
\item{Variable}{Description}
}
}
\usage{
out_list_demo
}
\description{
"Description in text"
}
\keyword{datasets}
