% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_cases}
\alias{sample_cases}
\title{Sample cases}
\format{
A vector of length 80.
\describe{
\item{Cases}{Numeric}
}
}
\usage{
sample_cases
}
\description{
Sample of aggregated case counts from a single location.
}
\keyword{datasets}
