% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_dates}
\alias{sample_dates}
\title{Sample dates}
\format{
A data.table data frame with X rows and Y variables:
\describe{
\item{Variable}{Description}
}
}
\usage{
sample_dates
}
\description{
"Description in text"
}
\keyword{datasets}
