% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_m_hier}
\alias{sample_m_hier}
\title{Sample hierachical model output}
\format{
An rstan object
\describe{
\item{list}{list}
}
}
\usage{
sample_m_hier
}
\description{
an example \code{rstan} output object from spatialrt()
}
\keyword{datasets}
