% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_multi_site}
\alias{sample_multi_site}
\title{Sample multi site}
\format{
A data.table data frame with X rows and Y variables:
\describe{
\item{Variable}{Description}
}
}
\usage{
sample_multi_site
}
\description{
"Description in text"
}
\keyword{datasets}
