% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_report_dates}
\alias{sample_report_dates}
\title{Sample report dates}
\format{
A vector of length 6380, one value per case.
\describe{
\item{Date}{Date of report}
}
}
\usage{
sample_report_dates
}
\description{
Line-list data, case report dates.
}
\keyword{datasets}
