% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialRt.R
\name{spatialRt}
\alias{spatialRt}
\title{Run Spatial R(t) estimation and Estimate Reproduction Numbers}
\usage{
spatialRt(report_dates, case_matrix, transfer_matrix, sip, v2 = FALSE, ...)
}
\arguments{
\item{report_dates}{A data frame ...}

\item{case_matrix}{A data frame ...}

\item{transfer_matrix}{A data frame ...}

\item{sip}{Vector of numeric values specifying the serial interval probabilities.}

\item{v2}{...}

\item{...}{Additional arguments passed to rstan::sampling()}
}
\value{
an object ...
}
\description{
This function performs ...
}
\details{
The function ensures ...
}
\examples{
\donttest{
data("sample_multi_site")
data("transfer_matrix")
Y <- as.matrix(sample_multi_site[, c(2, 3)])
for(i in 1:nrow(Y)) {
  for(j in 1:ncol(Y)) {
    Y[i,j] <- as.integer(Y[i,j])
  }
}
sip <- si(14, 4.29, 1.18, leading0 = FALSE)
sample_m_hier <- spatialRt(report_dates = sample_multi_site$date,
case_matrix = Y,
transfer_matrix = transfer_matrix,
v2 = FALSE,
sip = sip, chains = 1)
}
}
