% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transfer_matrix}
\alias{transfer_matrix}
\title{transfer_matrix}
\format{
A data.table data frame with X rows and Y variables:
\describe{
\item{Variable}{Description}
}
}
\usage{
transfer_matrix
}
\description{
"Description in text"
}
\keyword{datasets}
