% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{out_list_demo}
\alias{out_list_demo}
\title{Sample back-calculation output}
\format{
An backnow object
\describe{
\item{list}{list}
}
}
\usage{
out_list_demo
}
\description{
an example \code{backnow} output object from run_backnow()
}
\keyword{datasets}
