% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_backnow.R
\name{run_backnow}
\alias{run_backnow}
\title{Run Back Calculation and Estimate Reproduction Numbers}
\usage{
run_backnow(input, MAX_ITER, sip, NB_maxdelay, window_size, ...)
}
\arguments{
\item{input}{A data frame or list that includes epidemic data with either class 'caseCounts'
or 'lineList'. The input type determines initial processing steps.}

\item{MAX_ITER}{Integer, maximum number of iterations for the back-calculation model.
Requires at least 2000 iterations; high numbers can significantly increase runtime.}

\item{sip}{Vector of numeric values specifying the serial interval probabilities.}

\item{NB_maxdelay}{Integer, the maximum delay for the right-truncated negative binomial distribution used in modeling.}

\item{window_size}{Integer, the number of days of the R(t) averaging window.}

\item{...}{Additional arguments passed to underlying functions when converting input to the required format.}
}
\value{
an object of class \code{backnow} with the following structure

\if{html}{\out{<div class="sourceCode">}}\preformatted{ - est_back: back-calculated case counts

 - est_back_date: dates for back-calculated case counts

 - est_rt: back-calculated R(r)

 - est_rt_date: dates for back-calculated R(t)

 - report_date: a vector of dates, matches reported_cases

 - report_cases: a vector of reported cases

 - MAX_ITER: the input for `MAX_ITER`

 - NB_maxdelay: the input for `NB_maxdelay`

 - si: the input for serial interval `si`

 - window_size: the input for `window_size`
}\if{html}{\out{</div>}}
}
\description{
This function performs a back-calculation based on provided epidemic case count data,
estimating the time distribution of infections and reproduction numbers (r(t)).
It utilizes extensive input checks and parameter validation to ensure robust model execution.
}
\details{
The function ensures input data is of the correct class and processes it accordingly.
It handles different input classes by either converting \code{caseCounts} to \code{lineList} or
directly using \code{lineList}. The function stops with an error if the input doesn't meet expected standards.
It performs simulations to estimate both the back-calculation of initial infections and reproduction numbers
over time, while checking and adjusting for potential NA values and ensuring that all conditions for the
model parameters are met. Output includes estimates of initial infections and reproduction numbers along
with diagnostic statistics.
}
\examples{
\donttest{
data("sample_onset_dates")
data("sample_report_dates")
line_list <- create_linelist(sample_report_dates, sample_onset_dates)
sip <- si(14, 4.29, 1.18)
results <- run_backnow(
  line_list,
  MAX_ITER = as.integer(2000),
  sip = sip,
  NB_maxdelay = as.integer(20),
  window_size = as.integer(6))
}
}
