% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_dates}
\alias{sample_dates}
\title{Sample dates}
\format{
A vector of length 80.
\describe{
\item{Dates}{Dates of reported aggregated cases}
}
}
\usage{
sample_dates
}
\description{
Sample of case report dates from a single location.
}
\keyword{datasets}
