% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_location}
\alias{sample_location}
\title{Sample location}
\format{
A vector of length 80.
\describe{
\item{Location}{Character}
}
}
\usage{
sample_location
}
\description{
An vector of a single location to accompany \code{sample_dates} and \code{sample_cases}.
This is to emphasize that linelist functions are for a single location
}
\keyword{datasets}
