% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_multi_site}
\alias{sample_multi_site}
\title{Sample multi site}
\format{
A data.table data frame with 80 rows and 2 variables:
\describe{
\item{Cases for Site1}{Number of aggregated cases for Site 1}
\item{Cases for Site1}{Number of aggregated cases for Site 2}
}
}
\usage{
sample_multi_site
}
\description{
A matrix of daily aggregated cases from two sites (Hoth and Tatooine).
}
\keyword{datasets}
