% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_onset_dates}
\alias{sample_onset_dates}
\title{Sample onset dates}
\format{
A vector of length 6380, one value per case.
\describe{
\item{Date}{Date of onset}
}
}
\usage{
sample_onset_dates
}
\description{
Line-list data, onset dates, with some missing.
}
\keyword{datasets}
