% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialRt.R
\name{spatialRt}
\alias{spatialRt}
\title{Run Spatial R(t) estimation and Estimate Reproduction Numbers}
\usage{
spatialRt(report_dates, case_matrix, transfer_matrix, sip, v2 = FALSE, ...)
}
\arguments{
\item{report_dates}{A data frame ...}

\item{case_matrix}{A data frame ...}

\item{transfer_matrix}{A data frame ...}

\item{sip}{Vector of numeric values specifying the serial interval probabilities.}

\item{v2}{...}

\item{...}{Additional arguments passed to rstan::sampling()}
}
\value{
an object ...
}
\description{
This function performs ...
}
\details{
The function ensures ...
}
\examples{
\donttest{
print("duh")
}
}
