% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transfer_matrix}
\alias{transfer_matrix}
\title{Transfer matrix}
\format{
A data.table data frame with 160 rows and 2 variables:
\describe{
\item{Transfer to Site1}{Fraction of transfer to Site 1}
\item{Transfer to Site2}{Fraction of transfer to Site 2}
}
}
\usage{
transfer_matrix
}
\description{
A matrix of daily transfers between two sites (Hoth and Tatooine).
Each row sums to 1.
}
\keyword{datasets}
