% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhiteLabRt-package.R
\docType{package}
\name{WhiteLabRt-package}
\alias{WhiteLabRt-package}
\alias{WhiteLabRt}
\title{The 'WhiteLabRt' package.}
\description{
A collection of functions related to novel methods for
estimating reproduction number, R(t), created by the lab of
Professor Laura White at Boston University School of Public Health.

Currently implemented methods include (1) Temporal R(t) estimation:
Two-step Bayesian back and nowcasting for linelist data with missing
reporting delays, adapted in 'STAN' from
\href{https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1009210}{Li et. al. 2021},
and (2) Spatial R(t) estimation: Calculating time-varying reproduction number,
R(t), assuming a flux of infectors between various adjacent states,
in 'STAN' from \href{https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1010434}{Zhou et. al. 2021}.
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org
}
\author{
\strong{Maintainer}: Chad Milando \email{cmilando@bu.edu}

Other contributors:
\itemize{
  \item Tenglong Li [contributor]
  \item Zhenwei Zhou [contributor]
  \item Laura White [contributor]
}

}
